Refactoring: Separated feed parser
Moved feed parser out of the podcast module and created
its own module, for better code organization.
